+++ src/program/ui/InputEditorModel.cpp
@@ -109,25 +109,41 @@ QVariant InputEditorModel::data(const QModelIndex &index, int role) const
         QColor color = QGuiApplication::palette().window().color();
         int r, g, b;
         color.getRgb(&r, &g, &b, nullptr);
+        
         if (color.lightness() > 128) {
             /* Light theme */
             if (index.row() == static_cast<int>(context->framecount))
                 color.setRgb(r - 0x30, g - 0x10, b);
-            else if (index.row() < static_cast<int>(context->framecount))
-                color.setRgb(r - 0x30, g, b - 0x30);
-            else
-                color.setRgb(r, g, b - 0x18);
+            else if (index.row() < static_cast<int>(context->framecount)) {
+                if (movie->editor->isDraw(index.row()))
+                    color.setRgb(r - 0x30, g, b - 0x30);
+                else
+                    color.setRgb(r, g - 0x30, b - 0x30);
+            }
+            else {
+                if (movie->editor->isDraw(index.row()))
+                    color.setRgb(r, g, b - 0x18);
+                else
+                    color.setRgb(r, g - 0x18, b - 0x18);
+            }
         }
         else {
             /* Dark theme */
             if (index.row() == static_cast<int>(context->framecount))
                 color.setRgb(r, g + 0x10, b + 0x20);
-            else if (index.row() < static_cast<int>(context->framecount))
-                color.setRgb(r, g + 0x18, b);
-            else
-                color.setRgb(r + 0x08, g + 0x08, b);
+            else if (index.row() < static_cast<int>(context->framecount)) {
+                if (movie->editor->isDraw(index.row()))
+                    color.setRgb(r, g + 0x18, b);
+                else
+                    color.setRgb(r + 0x18, g, b);
+            }
+            else {
+                if (movie->editor->isDraw(index.row()))
+                    color.setRgb(r + 0x08, g + 0x08, b);
+                else
+                    color.setRgb(r + 0x08, g, b);
+            }
         }
-
 
         /* Frame column */
         if (index.column() <= 1) {