+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 * Can remove multiple ram watches
 * Implement greenzone using existing savestates (#134)
 * Add an option to allow games to warp the mouse pointer (#378)
+* Add nondraw frames in input editor
 
 ### Changed